feat: add configuration files for clang-tidy, editorconfig, and CI workflow; implement pre-commit hooks and update build scripts

+++ src/solve_worker.cpp
@@ -45,7 +45,7 @@ void SolveWorker::Execute()
 		on_progress = [tsfn](size_t done, size_t total) {
 			auto* value = new std::pair<size_t, size_t>(done, total);
 			auto const status = tsfn.NonBlockingCall(value, [](Napi::Env env, Napi::Function jsCallback, std::pair<size_t, size_t>* value) {
-				jsCallback.Call({ Napi::Number::New(env, value->first), Napi::Number::New(env, value->second) });
+				jsCallback.Call({ Napi::Number::New(env, static_cast<double>(value->first)), Napi::Number::New(env, static_cast<double>(value->second)) });
 				delete value;
 			});
 			if (status != napi_ok)
